abstract TableRow
extract TableRow component and associated logic from Profiles.tsx
